requirements.txt
langchain==1.2.10
langchain_community==0.4.1
langchain_core==1.2.13
langchain_experimental==0.4.1
langchain_google_genai==4.2.0
langchain_groq==1.1.2
langchain_huggingface==1.2.0
langchain_neo4j==0.8.0
langchain_tavily==0.2.17
langchain_text_splitters==1.1.0
langgraph==1.0.8
neo4j==6.1.0
pydantic==2.12.5
python-dotenv==1.2.1
PyYAML==6.0.3
setuptools==80.10.2
typing_extensions==4.15.0
sentence-transformers==5.2.2
rank_bm25==0.2.2
chromadb==0.5.0